Re added cheepservice
Added cheepservices back for onion layers.

Fixed razor pages and added the layer inbetween
